CLOUDP-187014: migrate quickstart and setup code
